Move latest/previous related state functions
Move the latest and previous version and migration functions into their
own file in the `pkg/state` package.
